python/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
mcp>=1.2
PyJWT[crypto]>=2.8
httpx>=0.27
bcrypt>=4.2
msgpack>=1.0
python-dotenv>=1.0
python-multipart>=0.0.9
